+++ samples/samples_cpp.cpp
@@ -1420,7 +1420,7 @@ int main( int argc, char *argv[] ) {
 	bool creditsEnd=false;
 	int cur_renderer=0;
 	static const char *renderer_name[TCOD_NB_RENDERERS] = {
-		"F1 GLSL   ","F2 OPENGL ","F3 SDL    "
+		"F1 GLSL   ","F2 OPENGL ","F3 SDL    ","F4 SDL2   "
 	};
 
 	// initialize the root console (open the game window)
@@ -1477,6 +1477,7 @@ int main( int argc, char *argv[] ) {
 	}
 
 	TCODConsole::initRoot(80,50,"libtcod C++ sample",fullscreen,renderer);
+	atexit(TCOD_quit);
 	do {
 		if (! creditsEnd) {
 			creditsEnd=TCODConsole::renderCredits(60,43,false);
@@ -1574,7 +1575,10 @@ int main( int argc, char *argv[] ) {
 			TCODSystem::setRenderer(TCOD_RENDERER_OPENGL);
 		} else if (key.vk==TCODK_F3) {
 			TCODSystem::setRenderer(TCOD_RENDERER_SDL);
+		} else if (key.vk==TCODK_F4) {
+			TCODSystem::setRenderer(TCOD_RENDERER_SDL2);
 		}
 	} while (!TCODConsole::isWindowClosed());
+	TCOD_quit();
 	return 0;
 }