+++ src/color/canvas.cc
@@ -17,7 +17,7 @@ Image load(const std::string& filename)
   Image image(img_width, img_height);
   std::vector<unsigned char>::iterator img_iter = img_data.begin();
   for (int y = 0; y < image.height(); ++y) {
-    for (int x = 0; y < image.width(); ++x) {
+    for (int x = 0; x < image.width(); ++x) {
       image.at(x, y) = ColorRGBA{
           img_iter[0], img_iter[1], img_iter[2], img_iter[3]};
       img_iter += 4;