+++ src/sdl2/sdl2_display.cc
@@ -19,25 +19,29 @@ WindowedDisplay::WindowedDisplay(std::pair<int, int> window_size,
   if (width < 0 || height < 0) {
     throw std::invalid_argument("width and height must be non-negative.");
   }
-  if (SDL_Init(SDL_INIT_VIDEO)) { throw std::runtime_error(SDL_GetError()); }
-  window_ = SDL_CreateWindow(
-      nullptr, SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED,
-      width, height, window_flags);
+  if (SDL_Init(SDL_INIT_VIDEO | SDL_INIT_EVENTS | SDL_INIT_TIMER)) {
+    throw std::runtime_error(SDL_GetError());
+  }
+  window_ = std::shared_ptr<SDL_Window>(
+      SDL_CreateWindow(
+          "Test", SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED,
+          width, height, window_flags),
+      [](SDL_Window* window){ SDL_DestroyWindow(window); });
   if (!window_) { throw std::runtime_error(SDL_GetError()); }
 }
 WindowedDisplay::~WindowedDisplay()
 {
-  if (window_) { SDL_DestroyWindow(window_); }
+  //if (window_) { SDL_DestroyWindow(window_); }
 }
 void WindowedDisplay::set_title(const std::string title)
 {
   if (!window_) { throw std::logic_error("Unresolved class invariant."); }
-  SDL_SetWindowTitle(window_, title.c_str());
+  SDL_SetWindowTitle(window_.get(), title.c_str());
 }
 std::string WindowedDisplay::get_title()
 {
   if (!window_) { throw std::logic_error("Unresolved class invariant."); }
-  return std::string(SDL_GetWindowTitle(window_));
+  return std::string(SDL_GetWindowTitle(window_.get()));
 }
 
 SDL2Display::SDL2Display(std::shared_ptr<Tileset> tileset,
@@ -45,14 +49,16 @@ SDL2Display::SDL2Display(std::shared_ptr<Tileset> tileset,
 : WindowedDisplay(window_size, window_flags)
 {
   // Configure SDL2 renderer.
-  renderer_ = SDL_CreateRenderer(get_window(), -1, SDL_RENDERER_TARGETTEXTURE);
+  renderer_ = std::shared_ptr<SDL_Renderer>(
+      SDL_CreateRenderer(get_window(), -1, SDL_RENDERER_TARGETTEXTURE),
+      [](SDL_Renderer* renderer){ SDL_DestroyRenderer(renderer); });
   if (!renderer_) { throw std::runtime_error(SDL_GetError()); }
   // Configure libtcod renderer.
   set_tileset(tileset);
 }
 SDL2Display::~SDL2Display()
 {
-  if (renderer_) { SDL_DestroyRenderer(renderer_); }
+  //if (renderer_) { SDL_DestroyRenderer(renderer_); }
 }
 void SDL2Display::set_tileset(std::shared_ptr<Tileset> tileset)
 {
@@ -60,14 +66,15 @@ void SDL2Display::set_tileset(std::shared_ptr<Tileset> tileset)
   if (!tileset) {
     throw std::invalid_argument("tileset must not be nullptr.");
   }
-  tcod_renderer_ = SDL2Renderer(renderer_, tileset);
+  tcod_renderer_ = SDL2Renderer(renderer_.get(), tileset);
 }
 void SDL2Display::present(const TCOD_Console* console)
 {
   if (!renderer_) { throw std::logic_error("Unresolved class invariant."); }
   SDL_Texture* backbuffer = tcod_renderer_.render(console);
-  SDL_RenderCopy(renderer_, backbuffer, nullptr, nullptr);
-  SDL_RenderPresent(renderer_);
+  SDL_RenderClear(renderer_.get());
+  SDL_RenderCopy(renderer_.get(), backbuffer, nullptr, nullptr);
+  SDL_RenderPresent(renderer_.get());
 }
 } // namespace sdl2
 } // namespace tcod