+++ src/sdl2/sdl2_display.h
@@ -3,6 +3,7 @@
 #define LIBTCOD_SDL2_SDL2_DISPLAY_H_
 
 #ifdef __cplusplus
+#include <memory>
 #include <utility>
 #endif // __cplusplus
 
@@ -23,9 +24,9 @@ class WindowedDisplay: public engine::Display {
   ~WindowedDisplay();
   virtual void set_title(const std::string title) override;
   virtual std::string get_title() override;
-  SDL_Window* get_window() { return window_; }
+  SDL_Window* get_window() { return window_.get(); }
  private:
-  SDL_Window* window_;
+  std::shared_ptr<SDL_Window> window_;
 };
 /**
  *  Display interface using a basic SDL2 renderer.
@@ -38,8 +39,7 @@ class SDL2Display: public WindowedDisplay {
   virtual void set_tileset(std::shared_ptr<Tileset> tileset) override;
   virtual void present(const TCOD_Console*) override;
  private:
-  SDL_Window* window_;
-  SDL_Renderer* renderer_;
+  std::shared_ptr<SDL_Renderer> renderer_;
   SDL2Renderer tcod_renderer_;
 };
 } // namespace sdl2